Twitter Stream Overhaul
Twitter only allow one stream per user. This meant if someone opened
the page somewhere else (another tab or computer) it would boot the
previous client off.

Now we have seperated the part of the code that turns on the twitter
stream and the part of the code that connects and recieves the data